fixed heap corruption
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@6577 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ clunk/object.cpp
@@ -62,6 +62,8 @@ void Object::cancel_all() {
 }
 
 Object::~Object() {
+	if (dead)
+		return;
 	AudioLocker l;
 	cancel_all();
 	context->delete_object(this);
